client/requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
APScheduler==3.9.1
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
blessed==1.17.6
bs4==0.0.1
cachetools==4.2.4
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cycler==0.11.0
discord.py==1.7.3
emoji==1.6.1
Flask==2.0.1
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fontawesomefree==6.1.2
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h5py==3.6.0
idna==2.10
importlib-metadata==4.10.0
inquirer==2.7.0
instaloader==4.7.4
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.1.0
Keras==2.1.6
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
libclang==12.0.0
lxml==4.6.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
multidict==5.1.0
multitasking==0.0.9
nltk==3.6.7
numpy==1.21.0
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.0
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pygame==2.0.1
PyJWT==2.3.0
pylint==2.6.0
pymongo==4.2.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
readchar==2.0.1
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
six==1.15.0
soupsieve==2.3.1
SQLAlchemy==1.4.25
termcolor==1.1.0
text2emotion==0.0.5
toml==0.10.2
tqdm==4.62.3
twilio==7.5.0
typing-extensions==3.10.0.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
yfinance==0.1.63
zipp==3.6.0
